+++ src/eval_framework/response_generator.py
@@ -1,3 +1,4 @@
+import logging
 import time
 import traceback
 from datetime import UTC, datetime
@@ -20,8 +21,9 @@ from eval_framework.tasks.base import Language, ResponseType, Sample
 from eval_framework.tasks.eval_config import EvalConfig
 from eval_framework.tasks.perturbation import create_perturbation_class
 from eval_framework.tasks.utils import raise_errors
-from eval_framework.utils.logging_config import get_logger
 from template_formatting.formatter import Message, Role
+
+logger = logging.getLogger(__name__)
 
 
 def map_language_to_value(
@@ -38,9 +40,6 @@ def map_language_to_value(
             return {k: (v[0].value, v[1].value) for k, v in language.items()}  # type: ignore[index]
     else:
         raise ValueError(f"Invalid language: {language}")
-
-
-logger = get_logger(__name__)
 
 
 class ResponseGenerator:
