+++ src/eval_framework/run.py
@@ -1,5 +1,6 @@
 import argparse
 import datetime
+import logging
 from pathlib import Path
 from typing import Any
 
@@ -11,9 +12,9 @@ except ImportError:
 from eval_framework.context.local import LocalContext
 from eval_framework.main import main
 from eval_framework.task_loader import load_extra_tasks
-from eval_framework.utils.logging_config import get_logger, setup_logging
+from eval_framework.utils import setup_logging
 
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 
 CONTEXT = {
     "local": LocalContext,
